# openapi.yaml
openapi: 3.0.0
info:
  title: RaftaarFreight API
  description: 'Pakistan''s premier real-time truck booking and delivery marketplace API'
  version: 1.0.0
  contact:
    name: RaftaarFreight Support
    email: support@raftaarfreight.com
  license:
    name: MIT
servers:
  - url: https://api.raftaarfreight.com
    description: Production
  - url: http://localhost:3001
    description: Development
...
paths:
...
  /api/bookings:
    get:
      tags:
        - Bookings
      summary: List bookings
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, in_progress, completed, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      tags:
        - Bookings
      summary: Create new booking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pickupLocation:
                  $ref: '#/components/schemas/Location'
                dropoffLocation:
                  $ref: '#/components/schemas/Location'
                pickupTime:
                  type: string
                  format: date-time
                quantity:
                  type: integer
                  minimum: 1
                weight:
                  type: number
                  minimum: 0
                description:
                  type: string
                truckType:
                  type: string
                  enum: [bike, auto, hatchback, sedan, suv, van, pickup, truck]
              required: [pickupLocation, dropoffLocation, pickupTime, quantity, weight]
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
  /api/bookings/{id}:
    get:
      tags:
        - Bookings
      summary: Get booking details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found

    patch:
      tags:
        - Bookings
      summary: Update booking
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [cancelled]
      responses:
        '200':
          description: Booking updated
        '403':
          description: Not authorized to update
...
components:
  schemas:
...
    Location:
      type: object
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
        address:
          type: string
        city:
          type: string
        province:
          type: string
      required: [lat, lng]
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
        driverId:
          type: string
          nullable: true
        pickupLocation:
          $ref: '#/components/schemas/Location'
        dropoffLocation:
          $ref: '#/components/schemas/Location'
        pickupTime:
          type: string
          format: date-time
        dropoffTime:
          type: string
          format: date-time
          nullable: true
        quantity:
          type: integer
        weight:
          type: number
        description:
          type: string
        status:
          type: string
          enum: [pending, accepted, in_progress, completed, cancelled]
        price:
          type: number
        paymentStatus:
          type: string
          enum: [pending, completed, failed, refunded]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...